+++ tests/conftest.py
@@ -2,7 +2,8 @@ import pytest
 
 from archie.auth.user_token import ARCGIS_ONLINE_BASE_URL, UserTokenAuth
 from archie.client import ArchieClient
-from tests.helpers import StaticTokenAuth
+from archie.services.base import BaseService
+from tests.helpers import BASE_URL, StaticTokenAuth, MinimalService
 
 
 @pytest.fixture
@@ -23,3 +24,14 @@ def client() -> ArchieClient:
         base_url="https://services.arcgis.com/instance",
         auth=StaticTokenAuth("test-token"),
     )
+
+
+@pytest.fixture
+def mock_client(mocker):
+    return mocker.create_autospec(ArchieClient)
+
+
+@pytest.fixture
+def service(mock_client) -> BaseService:
+    """A default BaseService instance for use in tests."""
+    return MinimalService(client=mock_client, url=BASE_URL)