+++ CHANGELOG.md
@@ -14,5 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Base async auth handler class called `ArcGISAuth` that mandates a `get_token()` method in any subclasses.
 - User token auth flows, where a user supplies a username and password in exchange for a token from ESRI's generateToken endpoint. Implemented as `UserTokenAuth`.
 - Base httpx client that handles authentication, enforces response formatting, and routes get and post requests from any endpoint. Implemented as `ArchieClient`.
+- Base service class that handles service metadata retreival, url validation, and client injection, implemented as `BaseService`.
 
 [Unreleased]: https://github.com/cityofboulder/archie